pyproject.toml
[project]
name = "resume-builder"
version = "0.1.0"
description = "Hybrid CLI + Claude Code plugin for resume generation, ATS scoring, JD matching, anti-fabrication, and session memory"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "reportlab>=4.0",
    "python-docx>=1.0",
    "jinja2>=3.1",
    "click>=8.0",
    "rich>=13.0",
    "mcp>=1.27.0",
]

[project.scripts]
rb = "resume_builder.cli:main"
rb-mcp = "resume_builder.mcp_server:main"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/resume_builder"]
